minerva-vault-api/requirements.txt
asgiref==3.8.1
Django==5.1.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.0
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2025.1
sqlparse==0.5.3
watchdog==3.0.0
